llm_search_template/requirements.txt
python-dotenv
flask
requests
streamlit
beautifulsoup4
google-cloud-aiplatform  # For accessing Gemini APIs
